feat: register and remove pix key in BCB
[tests present]

+++ src/main/kotlin/br/com/zup/service/RemovePixKeyService.kt
@@ -1,9 +1,12 @@
 package br.com.zup.service
 
+import br.com.zup.client.bcb.BcbClient
+import br.com.zup.client.bcb.model.request.BCBDeletePixKeyRequest
 import br.com.zup.grpc.exception.PixKeyNotFoundException
 import br.com.zup.model.domain.PixKey
 import br.com.zup.model.request.RemovePixKeyRequest
 import br.com.zup.repository.PixKeyRepository
+import io.micronaut.http.HttpStatus
 import io.micronaut.validation.Validated
 import java.util.*
 import javax.inject.Inject
@@ -14,7 +17,8 @@ import javax.validation.Valid
 @Validated
 @Singleton
 class RemovePixKeyService(
-    @Inject val pixKeyRepository: PixKeyRepository
+    @Inject val pixKeyRepository: PixKeyRepository,
+    @Inject val bcbClient: BcbClient
 ) {
 
     @Transactional
@@ -26,6 +30,13 @@ class RemovePixKeyService(
         ).orElseThrow {
             PixKeyNotFoundException("Chave não encontrada!")
         }
+
+        val deleteRequestBody = BCBDeletePixKeyRequest.of(key = pixKey.keyValue, participant = pixKey.linkedBankAccount.ispb)
+
+        val deleteResponse = bcbClient.deletePixKey(deleteRequestBody.key, deleteRequestBody)
+
+        if (deleteResponse.status != HttpStatus.OK)
+            throw IllegalStateException("Erro ao deletar chave PIX no Banco Central do Brasil (BCB)")
 
         pixKeyRepository.delete(pixKey)
